backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.10.3
python-docx>=1.1.0
fpdf2>=2.7.9
scikit-learn>=1.4.0
xgboost>=2.0.3
networkx>=3.2.1
pandas>=2.2.0
numpy>=1.26.4
groq>=0.9.0
python-dotenv>=1.0.0
